Refactor of getkey API
